move initial snapshot iterators creation to a new class
